re(re[re])organize the project
