+++ cli/src/makemkv_headless_cli/heuristics/HeuristicRegistry.py
