+++ cli/src/makemkv_headless_cli/heuristics/NamedHeuristic.py
